+++ projects/igniteui-angular/src/lib/chips/chip.spec.ts
@@ -1,4 +1,4 @@
-import { Component, ViewChild, ViewChildren, QueryList } from '@angular/core';
+﻿import { Component, ViewChild, ViewChildren, QueryList } from '@angular/core';
 import {
     async,
     TestBed
@@ -9,20 +9,59 @@ import { IgxIconModule } from '../icon/index';
 import { IgxChipsModule } from './chips.module';
 import { IgxChipComponent } from './chip.component';
 import { IgxChipsAreaComponent } from './chips-area.component';
+import { IgxPrefixDirective } from './../directives/prefix/prefix.directive';
+import { IgxConnectorDirective } from './connector.directive';
+import { IgxLabelDirective } from './../directives/label/label.directive';
+import { IgxSuffixDirective } from './../directives/suffix/suffix.directive';
+import { DisplayDensity } from 'dist/igniteui-angular/lib/core/utils';
 
 @Component({
     template: `
         <igx-chips-area #chipsArea>
             <igx-chip #chipElem *ngFor="let chip of chipList"
-            [id]="chip.id" [draggable]="chip.draggable" [removable]="chip.removable" [selectable]="chip.selectable">
-                <igx-icon igxPrefix fontSet="material" [name]="'drag_indicator'"></igx-icon>
+            [id]="chip.id" [draggable]="chip.draggable" [removable]="chip.removable" [selectable]="chip.selectable"
+            [displayDensity]="chip.density" (onRemove)="chipRemoved($event)">
                 <span #label [class]="'igx-chip__text'">{{chip.text}}</span>
                 <igx-icon class="igx-chip__dir-icon" igxConnector fontSet="material" [name]="'forward'"></igx-icon>
+                <igx-icon igxPrefix fontSet="material" [name]="'drag_indicator'"></igx-icon>
             </igx-chip>
         </igx-chips-area>
     `
 })
-export class TestChipComponent {
+class TestChipComponent {
+
+    public chipList = [
+        { id: 'Country', text: 'Country', removable: false, selectable: false, draggable: true },
+        { id: 'City', text: 'City', removable: true, selectable: true, draggable: true, density: 'comfortable' },
+        { id: 'Town', text: 'Town', removable: true, selectable: true, draggable: true, density: 'compact' },
+        { id: 'FirstName', text: 'First Name', removable: true , selectable: true, draggable: true, density: 'cosy' },
+    ];
+
+    @ViewChild('chipsArea', { read: IgxChipsAreaComponent})
+    public chipsArea: IgxChipsAreaComponent;
+
+    @ViewChildren('chipElem', { read: IgxChipComponent})
+    public chips: QueryList<IgxChipComponent>;
+
+    chipRemoved(event) {
+        this.chipList = this.chipList.filter((item) => {
+            return item.id !== event.owner.id;
+        });
+        this.chipsArea.cdr.detectChanges();
+    }
+}
+
+@Component({
+    template: `
+        <igx-chips-area>
+            <igx-chip *ngFor="let chip of chipList">
+                <span igxLabel>label</span>
+                <span igxSuffix>suf</span>
+            </igx-chip>
+        </igx-chips-area>
+    `
+})
+class TestChipsLabelAndSuffixComponent {
 
     public chipList = [
         { id: 'Country', text: 'Country', removable: false, selectable: false, draggable: true },
@@ -39,13 +78,33 @@ export class TestChipComponent {
 }
 
 describe('IgxChip', () => {
-    const CHIP_ITEM_AREA = 'igx-chip__item chip-area';
+    const CHIP_ITEM_AREA = 'igx-chip__item';
     const CHIP_CONNECTOR = 'igx-chip__connector';
+
+    function simulatePointerEvent(eventName: string, element, x, y) {
+        const options: PointerEventInit = {
+            view: window,
+            bubbles: true,
+            cancelable: true,
+            pointerId: 1
+        };
+        const pointerEvent = new PointerEvent(eventName, options);
+        Object.defineProperty(pointerEvent, 'pageX', { value: x, enumerable: true });
+        Object.defineProperty(pointerEvent, 'pageY', { value: y, enumerable: true });
+        return new Promise((resolve, reject) => {
+            element.dispatchEvent(pointerEvent);
+            resolve();
+        });
+    }
 
     beforeEach(async(() => {
         TestBed.configureTestingModule({
             declarations: [
-                TestChipComponent
+                TestChipComponent,
+                TestChipsLabelAndSuffixComponent,
+                IgxPrefixDirective,
+                IgxSuffixDirective,
+                IgxLabelDirective
             ],
             imports: [FormsModule, IgxIconModule, IgxChipsModule]
         }).compileComponents();
@@ -89,18 +148,16 @@ describe('IgxChip', () => {
         const fix = TestBed.createComponent(TestChipComponent);
         fix.detectChanges();
 
-        const chipElems = fix.debugElement.queryAll(By.directive(IgxChipComponent));
+        const chipAreaElem = fix.debugElement.queryAll(By.directive(IgxChipsAreaComponent));
+        const connectors = chipAreaElem[0].queryAll(By.directive(IgxConnectorDirective));
 
-        // The first 3 chips after each inner chip area should have a connector rendered.
-        expect(chipElems[0].nativeElement.children.length).toEqual(2);
-        expect(chipElems[0].nativeElement.children[1].className).toEqual(CHIP_CONNECTOR);
-        expect(chipElems[1].nativeElement.children.length).toEqual(2);
-        expect(chipElems[1].nativeElement.children[1].className).toEqual(CHIP_CONNECTOR);
-        expect(chipElems[2].nativeElement.children.length).toEqual(2);
-        expect(chipElems[2].nativeElement.children[1].className).toEqual(CHIP_CONNECTOR);
+        expect(connectors.length).toEqual(3);
 
         // The last 4th chip shouldn't have a connector.
-        expect(chipElems[3].nativeElement.children.length).toEqual(1);
+        const lastChipComponent = chipAreaElem[0].queryAll(By.directive(IgxChipComponent))[3];
+        const lastChipConnectors = lastChipComponent.queryAll(By.directive(IgxConnectorDirective));
+
+        expect(lastChipConnectors.length).toEqual(0);
     });
 
     it('should not trigger onRemove event when a chip is focused and delete button is pressed when not removable', () => {
@@ -143,5 +200,241 @@ describe('IgxChip', () => {
         fix.detectChanges();
 
         expect(secondChipComp.onRemove.emit).toHaveBeenCalled();
+    });
+
+    it('should set text in chips correctly', () => {
+        const fix = TestBed.createComponent(TestChipComponent);
+        fix.detectChanges();
+
+        const chipArea = fix.debugElement.queryAll(By.directive(IgxChipsAreaComponent));
+        const chipElements = chipArea[0].queryAll(By.directive(IgxChipComponent));
+        const firstChipTextElement = chipElements[0].queryAllNodes(By.css('.igx-chip__text'));
+        const firstChipText = firstChipTextElement[0].nativeNode.innerHTML;
+
+        expect(firstChipText).toContain('Country');
+
+        const secondChipTextElement = chipElements[1].queryAllNodes(By.css('.igx-chip__text'));
+        const secondChipText = secondChipTextElement[0].nativeNode.innerHTML;
+
+        expect(secondChipText).toContain('City');
+    });
+
+    it('should set chips label correctly', () => {
+        const fix = TestBed.createComponent(TestChipsLabelAndSuffixComponent);
+        fix.detectChanges();
+
+        const chipArea = fix.debugElement.queryAll(By.directive(IgxChipsAreaComponent));
+        const chipElements = chipArea[0].queryAll(By.directive(IgxChipComponent));
+        const firstChipLabel = chipElements[0].queryAll(By.directive(IgxLabelDirective));
+        const firstChipLabelText = firstChipLabel[0].nativeElement.innerHTML;
+
+        expect(firstChipLabelText).toEqual('label');
+    });
+
+    it('should set chips prefix correctly', () => {
+        const fix = TestBed.createComponent(TestChipComponent);
+        fix.detectChanges();
+
+        const chipArea = fix.debugElement.queryAll(By.directive(IgxChipsAreaComponent));
+        const chipElements = chipArea[0].queryAll(By.directive(IgxChipComponent));
+        const firstChipPrefix = chipElements[0].queryAll(By.directive(IgxPrefixDirective));
+        const firstChipIconName = firstChipPrefix[0].nativeElement.textContent;
+
+        expect(firstChipIconName).toContain('drag_indicator');
+    });
+
+    it('should set chips suffix correctly', () => {
+        const fix = TestBed.createComponent(TestChipsLabelAndSuffixComponent);
+        fix.detectChanges();
+
+        const chipArea = fix.debugElement.queryAll(By.directive(IgxChipsAreaComponent));
+        const chipElements = chipArea[0].queryAll(By.directive(IgxChipComponent));
+        const firstChipSuffix = chipElements[0].queryAll(By.directive(IgxSuffixDirective));
+        const firstChipSuffixText = firstChipSuffix[0].nativeElement.innerHTML;
+
+        expect(firstChipSuffixText).toEqual('suf');
+    });
+
+    it('should set chips suffix connector correctly', () => {
+        const fix = TestBed.createComponent(TestChipComponent);
+        fix.detectChanges();
+
+        const chipArea = fix.debugElement.queryAll(By.directive(IgxChipsAreaComponent));
+        const chipElements = chipArea[0].queryAll(By.directive(IgxChipComponent));
+        const firstChipSuffixConnector = chipElements[0].queryAll(By.directive(IgxConnectorDirective));
+        const firstChipSuffixConnectorIconName = firstChipSuffixConnector[0].nativeElement.textContent;
+
+        expect(firstChipSuffixConnectorIconName).toContain('forward');
+    });
+
+    it('should make chip comfortable when density is not set', () => {
+        const fix = TestBed.createComponent(TestChipComponent);
+        fix.detectChanges();
+
+        const components = fix.debugElement.queryAll(By.directive(IgxChipComponent));
+        const firstComponent = components[0];
+
+        expect(firstComponent.componentInstance.displayDensity).toEqual(DisplayDensity.comfortable);
+
+        // Assert default css class is applied
+        const comfortableComponents = fix.debugElement.queryAll(By.css('.igx-chip'));
+
+        expect(comfortableComponents.length).toEqual(2);
+        expect(comfortableComponents[0].nativeElement).toBe(firstComponent.nativeElement);
+    });
+
+    it('should make chip comfortable when density is set to comfortable', () => {
+        const fix = TestBed.createComponent(TestChipComponent);
+        fix.detectChanges();
+
+        const components = fix.debugElement.queryAll(By.directive(IgxChipComponent));
+        const secondComponent = components[1];
+
+        expect(secondComponent.componentInstance.displayDensity).toEqual(DisplayDensity.comfortable);
+
+        // Assert default css class is applied
+        const comfortableComponents = fix.debugElement.queryAll(By.css('.igx-chip'));
+
+        expect(comfortableComponents.length).toEqual(2);
+        expect(comfortableComponents[1].nativeElement).toBe(secondComponent.nativeElement);
+    });
+
+    it('should make chip compact when density is set to compact', () => {
+        const fix = TestBed.createComponent(TestChipComponent);
+        fix.detectChanges();
+
+        const components = fix.debugElement.queryAll(By.directive(IgxChipComponent));
+        const thirdComponent = components[2];
+
+        expect(thirdComponent.componentInstance.displayDensity).toEqual(DisplayDensity.compact);
+
+        // Assert compact css class is added
+        const compactComponents = fix.debugElement.queryAll(By.css('.igx-chip--compact'));
+
+        expect(compactComponents.length).toEqual(1);
+        expect(compactComponents[0].nativeElement).toBe(thirdComponent.nativeElement);
+    });
+
+    it('should make chip cosy when density is set to cosy', () => {
+        const fix = TestBed.createComponent(TestChipComponent);
+        fix.detectChanges();
+
+        const components = fix.debugElement.queryAll(By.directive(IgxChipComponent));
+        const fourthComponent = components[3];
+
+        expect(fourthComponent.componentInstance.displayDensity).toEqual(DisplayDensity.cosy);
+
+        // Assert cosy css class is added
+        const cosyComponents = fix.debugElement.queryAll(By.css('.igx-chip--cosy'));
+
+        expect(cosyComponents.length).toEqual(1);
+        expect(cosyComponents[0].nativeElement).toBe(fourthComponent.nativeElement);
+    });
+
+    it('should set correctly color of chip when color is set through code', () => {
+        const fix = TestBed.createComponent(TestChipComponent);
+        fix.detectChanges();
+        const chipColor = 'rgb(255, 0, 0)';
+
+        const components = fix.debugElement.queryAll(By.directive(IgxChipComponent));
+        const firstComponent = components[0];
+        const chipAreaElem = firstComponent.queryAll(By.css('.igx-chip__item'))[0];
+
+        firstComponent.componentInstance.color = chipColor;
+
+        expect(chipAreaElem.nativeElement.style.backgroundColor).toEqual(chipColor);
+        expect(firstComponent.componentInstance.color).toEqual(chipColor);
+    });
+
+    it('should delete chip when space button is pressed after chip delete button is focused', () => {
+        const spaceKeyEvent = new KeyboardEvent('keydown', {
+            'key': ' '
+        });
+
+        const fix = TestBed.createComponent(TestChipComponent);
+        fix.detectChanges();
+
+        let chipComponents = fix.debugElement.queryAll(By.directive(IgxChipComponent));
+
+        expect(chipComponents.length).toEqual(4);
+
+        const deleteButtonElement = fix.debugElement.queryAll(By.css('igx-icon.igx-chip__remove-icon'))[0];
+        deleteButtonElement.nativeElement.focus();
+
+        // Removes chip with id City, because country chip is unremovable
+        deleteButtonElement.nativeElement.dispatchEvent(spaceKeyEvent);
+        fix.detectChanges();
+
+        chipComponents = fix.debugElement.queryAll(By.directive(IgxChipComponent));
+        expect(chipComponents.length).toEqual(3);
+
+        const chipComponentsIds = fix.componentInstance.chipList.map(c => c.id);
+
+        expect(chipComponentsIds.length).toEqual(3);
+        expect(chipComponentsIds).not.toContain('City');
+    });
+
+    it('should delete chip when enter button is pressed after chip delete button is focused', () => {
+        const enterKeyEvent = new KeyboardEvent('keydown', {
+            'key': 'Enter'
+        });
+
+        const fix = TestBed.createComponent(TestChipComponent);
+        fix.detectChanges();
+        let chipComponents = fix.debugElement.queryAll(By.directive(IgxChipComponent));
+
+        expect(chipComponents.length).toEqual(4);
+
+        const deleteButtonElement = fix.debugElement.queryAll(By.css('igx-icon.igx-chip__remove-icon'))[0];
+        deleteButtonElement.nativeElement.focus();
+
+        // Removes chip with id City, because country chip is unremovable
+        deleteButtonElement.nativeElement.dispatchEvent(enterKeyEvent);
+        fix.detectChanges();
+
+        chipComponents = fix.debugElement.queryAll(By.directive(IgxChipComponent));
+        expect(chipComponents.length).toEqual(3);
+
+        const chipComponentsIds = fix.componentInstance.chipList.map(c => c.id);
+
+        expect(chipComponentsIds.length).toEqual(3);
+        expect(chipComponentsIds).not.toContain('City');
+    });
+
+    it('should affect the dragGhost density when chip has it set to compact', (done) => {
+        const fix = TestBed.createComponent(TestChipComponent);
+        fix.detectChanges();
+
+        const thirdChip = fix.componentInstance.chips.toArray()[2];
+        const thirdChipElem = thirdChip.chipArea.nativeElement;
+
+        const startingTop = thirdChipElem.getBoundingClientRect().top;
+        const startingLeft = thirdChipElem.getBoundingClientRect().left;
+        const startingBottom = thirdChipElem.getBoundingClientRect().bottom;
+        const startingRight = thirdChipElem.getBoundingClientRect().right;
+
+        const startingX = (startingLeft + startingRight) / 2;
+        const startingY = (startingTop + startingBottom) / 2;
+
+        simulatePointerEvent('pointerdown', thirdChipElem, startingX, startingY);
+        fix.detectChanges();
+
+        fix.whenStable().then(() => {
+            fix.detectChanges();
+            simulatePointerEvent('pointermove', thirdChipElem, startingX + 10, startingY + 10);
+
+            return fix.whenStable();
+        }).then(() => {
+            fix.detectChanges();
+
+            expect(thirdChip.dragDir['_dragGhost'].className).toEqual('igx-chip__item igx-chip__ghost--compact');
+
+            return fix.whenStable();
+        }).then(() => {
+            fix.detectChanges();
+            simulatePointerEvent('pointerup', thirdChip.dragDir['_dragGhost'], startingX + 10, startingY + 10);
+
+            done();
+        });
     });
 });