+++ projects/igniteui-angular/src/lib/directives/text-highlight/text-highlight.directive.ts
@@ -52,8 +52,19 @@ export class IgxTextHighlightDirective implements AfterViewInit, OnDestroy, OnCh
     @Input('groupName')
     public groupName = '';
 
+    private _value = '';
+
     @Input('value')
-    public value: any = '';
+    public get value(): any {
+        return this._value;
+    }
+    public set value(value: any) {
+        if (value === undefined || value === null) {
+            this._value = '';
+        } else {
+            this._value = value;
+        }
+    }
 
     @Input('row')
     public row: number;
@@ -180,6 +191,9 @@ export class IgxTextHighlightDirective implements AfterViewInit, OnDestroy, OnCh
                 this.clearChildElements(true);
                 this._lastSearchInfo.matchCount = this.getHighlightedText(text, caseSensitive);
             }
+        } else if (this._nodeWasRemoved) {
+            this._lastSearchInfo.searchedText = text;
+            this._lastSearchInfo.caseSensitive = caseSensitiveResolved;
         }
 
         return this._lastSearchInfo.matchCount;
@@ -212,7 +226,7 @@ export class IgxTextHighlightDirective implements AfterViewInit, OnDestroy, OnCh
 
             const elementToActivate = spans[index];
             this.renderer.addClass(elementToActivate, this.activeCssClass);
-            this.renderer.setAttribute(elementToActivate, 'style', 'background:orange;font-weight:bold');
+            this.renderer.setAttribute(elementToActivate, 'style', 'background:orange;font-weight:bold;color:black');
         }
     }
 
@@ -230,7 +244,7 @@ export class IgxTextHighlightDirective implements AfterViewInit, OnDestroy, OnCh
 
         const elementToDeactivate = spans[this._activeElementIndex];
         this.renderer.removeClass(elementToDeactivate, this.activeCssClass);
-        this.renderer.setAttribute(elementToDeactivate, 'style', 'background:yellow;font-weight:bold');
+        this.renderer.setAttribute(elementToDeactivate, 'style', 'background:yellow;font-weight:bold;color:black');
         this._activeElementIndex = -1;
     }
 
@@ -264,7 +278,7 @@ export class IgxTextHighlightDirective implements AfterViewInit, OnDestroy, OnCh
 
             this.appendText(stringValue.substring(previousMatchEnd, start));
             // tslint:disable-next-line:max-line-length
-            this.appendSpan(`<span class="${this.cssClass}" style="background:yellow;font-weight:bold">${stringValue.substring(start, end)}</span>`);
+            this.appendSpan(`<span class="${this.cssClass}" style="background:yellow;font-weight:bold;color:black">${stringValue.substring(start, end)}</span>`);
 
             previousMatchEnd = end;
             matchCount++;