+++ projects/igniteui-angular/src/lib/grid/grid-header.component.ts
@@ -36,10 +36,11 @@ export class IgxGridHeaderComponent implements OnInit, DoCheck, AfterViewInit {
 
     @HostBinding('class')
     get styleClasses() {
-        if (this.column.columnGroup) {
-            return `${this.column.headerClasses}`;
-        }
-        return `igx-grid__th ${this.column.headerClasses}`;
+        return[
+            this.column.columnGroup ? '' : `${this.defaultCssClass}`,
+            this.column.headerClasses,
+            this.column.dataType === DataType.Number ? `${this.numberCssClass}` : ''
+        ].join(' ');
     }
 
     @HostBinding('style.min-width')
@@ -74,11 +75,6 @@ export class IgxGridHeaderComponent implements OnInit, DoCheck, AfterViewInit {
             return this.sortDirection === SortingDirection.Asc ? 'arrow_upward' : 'arrow_downward';
         }
         return 'none';
-    }
-
-    @HostBinding('class.igx-grid__th--number')
-    get columnType() {
-        return this.column.dataType === DataType.Number;
     }
 
     @HostBinding('class.igx-grid__th--sorted')
@@ -120,6 +116,9 @@ export class IgxGridHeaderComponent implements OnInit, DoCheck, AfterViewInit {
     public resizeEndTimeout = /Firefox[\/\s](\d+\.\d+)/.test(navigator.userAgent) ? 200 : 0;
 
     protected sortDirection = SortingDirection.None;
+    protected defaultCssClass = 'igx-grid__th';
+    protected numberCssClass = 'igx-grid__th--number';
+
     private _startResizePos;
     private _pinnedMaxWidth;
 
@@ -152,22 +151,16 @@ export class IgxGridHeaderComponent implements OnInit, DoCheck, AfterViewInit {
 
     @HostListener('click', ['$event'])
     public onClick(event) {
-
         if (!this.column.grid.isColumnResizing) {
             event.stopPropagation();
             if (this.column.sortable) {
-                const grid = this.gridAPI.get(this.gridID);
-                const editableCell = this.gridAPI.get_cell_inEditMode(this.gridID);
-                if (editableCell) {
-                    this.gridAPI.escape_editMode(this.gridID, editableCell.cellID);
-                }
-                const groupingExpr = grid.groupingExpressions.find((expr) => expr.fieldName === this.column.field);
+                const groupingExpr = this.grid.groupingExpressions.find((expr) => expr.fieldName === this.column.field);
                 const sortDir = groupingExpr ?
                     this.sortDirection + 1 > SortingDirection.Desc ? SortingDirection.Asc  : SortingDirection.Desc
                     : this.sortDirection + 1 > SortingDirection.Desc ? SortingDirection.None : this.sortDirection + 1;
                 this.sortDirection = sortDir;
-                this.gridAPI.sort(this.gridID, this.column.field, this.sortDirection, this.column.sortingIgnoreCase);
-                grid.onSortingDone.emit({
+                this.grid.sort({ fieldName: this.column.field, dir: this.sortDirection, ignoreCase: this.column.sortingIgnoreCase});
+                this.grid.onSortingDone.emit({
                     dir: this.sortDirection,
                     fieldName: this.column.field,
                     ignoreCase: this.column.sortingIgnoreCase